+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/ImpTest.hs
@@ -212,9 +212,10 @@ import qualified Data.Text as T
 import Data.Tree
 import qualified GHC.Exts as GHC (fromList)
 import Lens.Micro
+import Prettyprinter (align, hsep, viaShow, vsep)
 import Test.Cardano.Ledger.Babbage.ImpTest
 import Test.Cardano.Ledger.Conway.Arbitrary ()
-import Test.Cardano.Ledger.Conway.TreeDiff ()
+import Test.Cardano.Ledger.Conway.TreeDiff (tableDoc)
 import Test.Cardano.Ledger.Core.KeyPair (KeyPair (..), mkCred)
 import Test.Cardano.Ledger.Core.Rational (IsRatio (..))
 import Test.Cardano.Ledger.Imp.Common
@@ -881,10 +882,10 @@ getEnactState = mkEnactState <$> getsNES (nesEsL . epochStateGovStateL)
 getProposals :: ConwayEraGov era => ImpTestM era (Proposals era)
 getProposals = getsNES $ nesEsL . esLStateL . lsUTxOStateL . utxosGovStateL . proposalsGovStateL
 
-logProposalsForest :: ConwayEraGov era => ImpTestM era ()
+logProposalsForest :: (ConwayEraGov era, HasCallStack) => ImpTestM era ()
 logProposalsForest = do
   proposals <- getProposals
-  logEntry $ proposalsShowDebug proposals True
+  logDoc $ proposalsShowDebug proposals True
 
 getCommitteeMembers ::
   ConwayEraImp era =>
@@ -949,11 +950,14 @@ getParameterChangeProposals = do
   pure $ ps ^. pGraphL . grPParamUpdateL . pGraphNodesL
 
 logProposalsForestDiff ::
-  (Era era, ToExpr (PParamsHKD StrictMaybe era)) =>
+  ( Era era
+  , ToExpr (PParamsHKD StrictMaybe era)
+  , HasCallStack
+  ) =>
   Proposals era ->
   Proposals era ->
   ImpTestM era ()
-logProposalsForestDiff pf1 pf2 = logEntry $ unlines ["Proposals Forest Diff:", diffExpr pf1 pf2]
+logProposalsForestDiff pf1 pf2 = logDoc $ vsep ["Proposals Forest Diff:", diffExpr pf1 pf2]
 
 -- | Looks up the governance action state corresponding to the governance action id
 lookupGovActionState ::
@@ -988,7 +992,7 @@ expectMissingGovActionId govActionId =
   impAnn "Expecting for gov action state to be missing" $ do
     lookupGovActionState govActionId >>= \case
       Just _ ->
-        expectationFailure $ "Found gov action state for govActionId: " <> show govActionId
+        expectationFailure $ "Found gov action state for govActionId: " <> ansiExprString govActionId
       Nothing -> pure ()
 
 -- | Builds a RatifyEnv from the current state
@@ -1111,13 +1115,12 @@ logAcceptedRatio aId = do
   dRepRatio <- calculateDRepAcceptedRatio aId
   committeeRatio <- calculateCommitteeAcceptedRatio aId
   spoRatio <- calculatePoolAcceptedRatio aId
-  logEntry $
-    unlines
-      [ ""
-      , "----- ACCEPTED RATIOS -----"
-      , "DRep accepted ratio:\t\t" <> show dRepRatio
-      , "Committee accepted ratio:\t" <> show committeeRatio
-      , "SPO accepted ratio:\t\t" <> show spoRatio
+  logDoc $
+    tableDoc
+      (Just "ACCEPTED RATIOS")
+      [ ("DRep accepted ratio:", viaShow dRepRatio)
+      , ("Committee accepted ratio:", viaShow committeeRatio)
+      , ("SPO accepted ratio:", viaShow spoRatio)
       ]
 
 getRatifyEnvAndState :: ConwayEraGov era => ImpTestM era (RatifyEnv era, RatifyState era)
@@ -1164,7 +1167,7 @@ isCommitteeAccepted gaId = do
 
 -- | Logs the results of each check required to make the governance action pass
 logRatificationChecks ::
-  (ConwayEraGov era, ConwayEraPParams era) =>
+  (ConwayEraGov era, ConwayEraPParams era, HasCallStack) =>
   GovActionId (EraCrypto era) ->
   ImpTestM era ()
 logRatificationChecks gaId = do
@@ -1180,35 +1183,49 @@ logRatificationChecks gaId = do
     members = foldMap' committeeMembers committee
     committeeState = reCommitteeState ratEnv
   curPParams <- getsNES $ nesEsL . epochStateGovStateL . curPParamsGovStateL
-  logEntry $
-    unlines
-      [ "----- RATIFICATION CHECKS -----"
-      , "prevActionAsExpected:\t" <> show (prevActionAsExpected gas ensPrevGovActionIds)
-      , "validCommitteeTerm:\t" <> show (validCommitteeTerm govAction curPParams currentEpoch)
-      , "notDelayed:\t\t??"
-      , "withdrawalCanWithdraw:\t" <> show (withdrawalCanWithdraw govAction curTreasury)
-      , "committeeAccepted:\t"
-          <> show (committeeAccepted ratEnv ratSt gas)
-          <> " [ To Pass: "
-          <> show (committeeAcceptedRatio members gasCommitteeVotes committeeState currentEpoch)
-          <> " >= "
-          <> show (votingCommitteeThreshold reCurrentEpoch ratSt committeeState (gasAction gas))
-          <> " ]"
-      , "spoAccepted:\t\t"
-          <> show (spoAccepted ratEnv ratSt gas)
-          <> " [ To Pass: "
-          <> show (spoAcceptedRatio ratEnv gas)
-          <> " >= "
-          <> show (votingStakePoolThreshold ratSt (gasAction gas))
-          <> " ]"
-      , "dRepAccepted:\t\t"
-          <> show (dRepAccepted ratEnv ratSt gas)
-          <> " [ To Pass: "
-          <> show (dRepAcceptedRatio ratEnv gasDRepVotes (gasAction gas))
-          <> " >= "
-          <> show (votingDRepThreshold ratSt (gasAction gas))
-          <> " ]"
-      , ""
+  logDoc $
+    tableDoc
+      (Just "RATIFICATION CHECKS")
+      [ ("prevActionAsExpected:", viaShow $ prevActionAsExpected gas ensPrevGovActionIds)
+      , ("validCommitteeTerm:", viaShow $ validCommitteeTerm govAction curPParams currentEpoch)
+      , ("notDelayed:", "??")
+      , ("withdrawalCanWithdraw:", viaShow $ withdrawalCanWithdraw govAction curTreasury)
+      ,
+        ( "committeeAccepted:"
+        , hsep
+            [ viaShow $ committeeAccepted ratEnv ratSt gas
+            , "["
+            , "To Pass:"
+            , viaShow $ committeeAcceptedRatio members gasCommitteeVotes committeeState currentEpoch
+            , ">="
+            , viaShow $ votingCommitteeThreshold reCurrentEpoch ratSt committeeState (gasAction gas)
+            , "]"
+            ]
+        )
+      ,
+        ( "spoAccepted:"
+        , hsep
+            [ viaShow $ spoAccepted ratEnv ratSt gas
+            , "["
+            , "To Pass:"
+            , viaShow $ spoAcceptedRatio ratEnv gas
+            , ">="
+            , viaShow $ votingStakePoolThreshold ratSt (gasAction gas)
+            , "]"
+            ]
+        )
+      ,
+        ( "dRepAccepted:"
+        , hsep
+            [ viaShow $ dRepAccepted ratEnv ratSt gas
+            , "["
+            , "To Pass:"
+            , viaShow $ dRepAcceptedRatio ratEnv gasDRepVotes (gasAction gas)
+            , ">="
+            , viaShow $ votingDRepThreshold ratSt (gasAction gas)
+            , "]"
+            ]
+        )
       ]
 
 -- | Submits a transaction that registers a hot key for the given cold key.
@@ -1293,11 +1310,11 @@ electBasicCommittee ::
     , GovPurposeId 'CommitteePurpose era
     )
 electBasicCommittee = do
-  logEntry "Setting up a DRep"
+  logString "Setting up a DRep"
   (drep, _, _) <- setupSingleDRep 1_000_000
   (spoC, _, _) <- setupPoolWithStake $ Coin 1_000_000
 
-  logEntry "Registering committee member"
+  logString "Registering committee member"
   coldCommitteeC <- KeyHashObj <$> freshKeyHash
   startEpochNo <- getsNES nesELL
   let
@@ -1322,49 +1339,54 @@ electBasicCommittee = do
   hotCommitteeC <- registerCommitteeHotKey coldCommitteeC
   pure (drep, hotCommitteeC, GovPurposeId gaidCommitteeProp)
 
-logCurPParams :: (EraGov era, ToExpr (PParamsHKD Identity era)) => ImpTestM era ()
+logCurPParams ::
+  ( EraGov era
+  , ToExpr (PParamsHKD Identity era)
+  , HasCallStack
+  ) =>
+  ImpTestM era ()
 logCurPParams = do
   pp <- getsNES $ nesEsL . curPParamsEpochStateL
-  logEntry $
-    unlines
+  logDoc $
+    vsep
       [ ""
       , "----- Current PParams -----"
-      , showExpr pp
+      , ansiExpr pp
       , "---------------------------"
       , ""
       ]
 
-proposalsShowDebug :: Era era => Proposals era -> Bool -> String
+proposalsShowDebug :: Era era => Proposals era -> Bool -> Doc AnsiStyle
 proposalsShowDebug ps showRoots =
-  unlines $
+  align . vsep $
     [ ""
     , "----- Proposals -----"
     , "Size"
-    , show $ proposalsSize ps
+    , viaShow $ proposalsSize ps
     , "OMap"
-    , show $ proposalsIds ps
+    , viaShow $ proposalsIds ps
     , ""
     , "Roots"
     , "> PParamUpdate"
-    , show $ ps ^. pRootsL . grPParamUpdateL
+    , viaShow $ ps ^. pRootsL . grPParamUpdateL
     , "> HardFork"
-    , show $ ps ^. pRootsL . grHardForkL
+    , viaShow $ ps ^. pRootsL . grHardForkL
     , "> Committee"
-    , show $ ps ^. pRootsL . grCommitteeL
+    , viaShow $ ps ^. pRootsL . grCommitteeL
     , "> Constitution"
-    , show $ ps ^. pRootsL . grConstitutionL
+    , viaShow $ ps ^. pRootsL . grConstitutionL
     ]
       <> ( if showRoots
             then
               [ "Hierarchy"
               , ">> PParamUpdate"
-              , show $ ps ^. pGraphL . grPParamUpdateL . pGraphNodesL
+              , viaShow $ ps ^. pGraphL . grPParamUpdateL . pGraphNodesL
               , ">> HardFork"
-              , show $ ps ^. pGraphL . grHardForkL . pGraphNodesL
+              , viaShow $ ps ^. pGraphL . grHardForkL . pGraphNodesL
               , ">> Committee"
-              , show $ ps ^. pGraphL . grCommitteeL . pGraphNodesL
+              , viaShow $ ps ^. pGraphL . grCommitteeL . pGraphNodesL
               , ">> Constitution"
-              , show $ ps ^. pGraphL . grConstitutionL . pGraphNodesL
+              , viaShow $ ps ^. pGraphL . grConstitutionL . pGraphNodesL
               ]
             else mempty
          )