+++ libs/cardano-ledger-binary/testlib/Test/Cardano/Ledger/Binary/TreeDiff.hs
@@ -8,9 +8,13 @@ module Test.Cardano.Ledger.Binary.TreeDiff (
   CBORBytes (..),
   HexBytes (..),
   showExpr,
+  ansiExpr,
+  ansiExprString,
   diffExpr,
+  diffExprString,
   diffExprCompact,
-  diffExprNoColor,
+  diffExprCompactString,
+  ansiDocToString,
   hexByteStringExpr,
   showHexBytesGrouped,
   expectExprEqual,
@@ -19,7 +23,10 @@ module Test.Cardano.Ledger.Binary.TreeDiff (
   Expr (App, Rec, Lst),
   defaultExprViaShow,
   trimExprViaShow,
+  tableDoc,
   Pretty (..),
+  Doc,
+  AnsiStyle,
   ansiWlPretty,
   ppEditExpr,
   ediff,
@@ -33,7 +40,6 @@ import Cardano.Crypto.Hash.Class ()
 import Cardano.Ledger.Binary
 import qualified Codec.CBOR.Read as CBOR
 import qualified Codec.CBOR.Term as CBOR
-import Control.Monad (unless)
 import Data.Bifunctor (bimap)
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Base16 as Base16
@@ -43,7 +49,12 @@ import Data.Foldable (toList)
 import Data.IP (IPv4, IPv6)
 import Data.Maybe.Strict (StrictMaybe)
 import Data.Sequence.Strict (StrictSeq)
+import qualified Data.Text.Lazy as TL
 import Data.TreeDiff
+import Prettyprinter (Doc)
+import qualified Prettyprinter as Pretty
+import Prettyprinter.Render.Terminal (AnsiStyle)
+import qualified Prettyprinter.Render.Terminal as Pretty
 import Test.Cardano.Slotting.TreeDiff ()
 import Test.Hspec (Expectation, HasCallStack, expectationFailure)
 import Test.Tasty.HUnit (Assertion, assertFailure)
@@ -53,6 +64,14 @@ import Test.Tasty.HUnit (Assertion, assertFailure)
 
 trimExprViaShow :: Show a => Int -> a -> Expr
 trimExprViaShow _n x = defaultExprViaShow x -- App (take n (drop 1 (show x)) ++ "..") []
+
+tableDoc :: Maybe (Doc AnsiStyle) -> [(String, Doc AnsiStyle)] -> Doc AnsiStyle
+tableDoc mTitle rows =
+  let w = foldr (max . length . fst) 0 rows
+      t = case mTitle of
+        Just title -> Pretty.hsep ["-----", title, "-----"] <> Pretty.line
+        Nothing -> mempty
+   in t <> Pretty.vsep [Pretty.fill (w + 1) (Pretty.pretty l) <> r | (l, r) <- rows]
 
 -- ===========================================================
 -- Orphan instances from external imports
@@ -82,16 +101,28 @@ instance ToExpr a => ToExpr (Sized a)
 --------------------------------------------------------------------------------
 
 showExpr :: ToExpr a => a -> String
-showExpr = show . ansiWlExpr . toExpr
+showExpr = show . prettyExpr . toExpr
 
-diffExpr :: ToExpr a => a -> a -> String
-diffExpr x y = show (ansiWlEditExpr (ediff x y))
+ansiExpr :: ToExpr a => a -> Doc AnsiStyle
+ansiExpr = ansiWlExpr . toExpr
 
-diffExprNoColor :: ToExpr a => a -> a -> String
-diffExprNoColor x y = show (prettyEditExpr (ediff x y))
+ansiExprString :: ToExpr a => a -> String
+ansiExprString = ansiDocToString . ansiExpr
 
-diffExprCompact :: ToExpr a => a -> a -> String
-diffExprCompact x y = show (ansiWlEditExprCompact (ediff x y))
+diffExpr :: ToExpr a => a -> a -> Doc AnsiStyle
+diffExpr x y = ansiWlEditExpr (ediff x y)
+
+diffExprString :: ToExpr a => a -> a -> String
+diffExprString x y = ansiDocToString $ diffExpr x y
+
+diffExprCompact :: ToExpr a => a -> a -> Doc AnsiStyle
+diffExprCompact x y = ansiWlEditExprCompact (ediff x y)
+
+diffExprCompactString :: ToExpr a => a -> a -> String
+diffExprCompactString x y = ansiDocToString $ diffExprCompact x y
+
+ansiDocToString :: Doc AnsiStyle -> String
+ansiDocToString = TL.unpack . Pretty.renderLazy . Pretty.layoutPretty Pretty.defaultLayoutOptions
 
 -- | Wraps regular ByteString, but shows and diffs it as hex
 newtype HexBytes = HexBytes {unHexBytes :: BS.ByteString}
@@ -187,15 +218,17 @@ expectExprEqual :: (Eq a, ToExpr a) => a -> a -> Expectation
 expectExprEqual = expectExprEqualWithMessage "Expected two values to be equal:"
 
 -- | Use this with HSpec, but with Tasty use 'assertExprEqualWithMessage' below
-expectExprEqualWithMessage :: (ToExpr a, Eq a, HasCallStack) => [Char] -> a -> a -> Expectation
-expectExprEqualWithMessage message expected actual =
-  unless (actual == expected) (expectationFailure msg)
-  where
-    msg = (if null message then "" else message ++ "\n") ++ diffExpr expected actual
+expectExprEqualWithMessage :: (ToExpr a, Eq a, HasCallStack) => String -> a -> a -> Expectation
+expectExprEqualWithMessage = requireExprEqualWithMessage (expectationFailure . ansiDocToString) . Pretty.pretty
 
 -- | Use this with Tasty, but with HSpec use 'expectExprEqualWithMessage' above
-assertExprEqualWithMessage :: (ToExpr a, Eq a, HasCallStack) => [Char] -> a -> a -> Assertion
-assertExprEqualWithMessage message expected actual =
-  unless (actual == expected) (assertFailure msg)
+assertExprEqualWithMessage :: (ToExpr a, Eq a, HasCallStack) => String -> a -> a -> Assertion
+assertExprEqualWithMessage = requireExprEqualWithMessage (assertFailure . ansiDocToString) . Pretty.pretty
+
+requireExprEqualWithMessage ::
+  (ToExpr a, Eq a, Monoid b) => (Doc AnsiStyle -> b) -> Doc AnsiStyle -> a -> a -> b
+requireExprEqualWithMessage fail_ message expected actual =
+  if actual == expected then mempty else fail_ doc
   where
-    msg = (if null message then "" else message ++ "\n") ++ diffExpr expected actual
+    doc = Pretty.width message (\w -> if w == 0 then diff else Pretty.line <> Pretty.indent 2 diff)
+    diff = diffExpr expected actual