+++ eras/conway/impl/CHANGELOG.md
@@ -15,12 +15,12 @@
   * `reCurrentEpochL`
   * `reCommitteeStateL`
 * Add a new field to `GovInfoEvent`
+* Changed return type of `proposalsShowDebug`
 
 ### `testlib`
 
-  * Removed `redelegateDRep` from `ImpTest`
-  * Changed signature of `delegateToDRep` to take a `Credential` parameter
-
+* Removed `redelegateDRep` from `ImpTest`
+* Changed signature of `delegateToDRep` to take a `Credential` parameter
 
 ## 1.16.1.0